# Optimize PropertyBag for reference types & code cleanup