--- a/src/Anixe.TransactionSteps/Anixe.TransactionSteps.csproj
+++ b/src/Anixe.TransactionSteps/Anixe.TransactionSteps.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
-    <Version>2.0.1</Version>
+    <Version>2.0.2</Version>
     <Title>Transaction Steps</Title>
     <Authors>WM</Authors>
   </PropertyGroup>