--- a/src/Anixe.TransactionSteps/IPropertyBag.cs
+++ b/src/Anixe.TransactionSteps/IPropertyBag.cs
@@ -11,5 +11,5 @@ namespace Anixe.TransactionSteps
     void Unset<T>();
     IPropertyBag Clone();
     IPropertyBag Clone(System.Type[] exclude);
-  }  
+  }
 }