--- a/src/Anixe.TransactionSteps/IStep.cs
+++ b/src/Anixe.TransactionSteps/IStep.cs
@@ -13,13 +13,13 @@ namespace Anixe.TransactionSteps
   public interface IStep
   {
     bool CanProcess();
-    bool IsAsync();    
+    bool IsAsync();
     Task ProcessAsync(CancellationToken token);
     void Process();
     IServiceProvider Services { get; set; }
     string Name { get; set; }
-    double TimeTaken { get; set; }    
-    int ProcessedItemsCount { get; set; }    
+    double TimeTaken { get; set; }
+    int ProcessedItemsCount { get; set; }
     LinkedList<IStep> Neighbourood { get; set; }
     LinkedListNode<IStep> Current { get; set; }
     bool WasFired { get; set; }