--- a/src/Anixe.TransactionSteps/IStepIterator.cs
+++ b/src/Anixe.TransactionSteps/IStepIterator.cs
@@ -8,16 +8,16 @@ namespace Anixe.TransactionSteps
   public interface IStepIterator<T> where T : IPropertyBag
   {
     List<StepStat> Stats { get; }
+
     Task<T> IterateAllAsync(
-      IServiceProvider services, 
-      LinkedList<IStep> steps, 
-      IStep errorHandler, 
+      IServiceProvider services,
+      LinkedList<IStep> steps,
+      IStep errorHandler,
       CancellationToken token);
 
     void IterateAll(
-      IServiceProvider services, 
-      LinkedList<IStep> steps, 
+      IServiceProvider services,
+      LinkedList<IStep> steps,
       IStep errorHandler);
-      
   }
 }