--- a/src/Anixe.TransactionSteps/IValueTaskStepIterator.cs
+++ b/src/Anixe.TransactionSteps/IValueTaskStepIterator.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;