--- a/src/Anixe.TransactionSteps/Predefined/AsyncStep.cs
+++ b/src/Anixe.TransactionSteps/Predefined/AsyncStep.cs
@@ -6,38 +6,37 @@ namespace Anixe.TransactionSteps.Predefined
 {
   public class AsyncStep : StepBase, IStep
   {
-    private Func<IStep, Task> action;
-    private Predicate<IStep> canProcessPredicate;
-    
+    private readonly Func<IStep, Task> action;
+    private readonly Predicate<IStep> canProcessPredicate;
 
     public AsyncStep(Func<IStep, Task> action, Predicate<IStep> canProcessPredicate = null)
     {
-      this.action = action;        
-      this.canProcessPredicate = canProcessPredicate ?? True;      
+      this.action = action;
+      this.canProcessPredicate = canProcessPredicate ?? True;
     }
-    
+
     public bool CanProcess()
     {
-        return this.canProcessPredicate(this);
+      return this.canProcessPredicate(this);
     }
 
     public bool IsAsync()
     {
-        return true;
+      return true;
     }
 
     public void Process()
     {
-      throw new NotImplementedException();
+      throw new NotSupportedException();
     }
 
     public async Task ProcessAsync(CancellationToken token)
     {
-      if(token.IsCancellationRequested)
+      if (token.IsCancellationRequested)
       {
         var tcs = new TaskCompletionSource<int>();
         tcs.SetCanceled();
-        await tcs.Task;        
+        await tcs.Task;
       }
       await this.action(this);
     }
@@ -45,6 +44,6 @@ namespace Anixe.TransactionSteps.Predefined
     private bool True(IStep step)
     {
       return true;
-    }        
+    }
   }
 }