--- a/src/Anixe.TransactionSteps/Predefined/Step.cs
+++ b/src/Anixe.TransactionSteps/Predefined/Step.cs
@@ -6,47 +6,44 @@ namespace Anixe.TransactionSteps.Predefined
 {
   public class Step : StepBase, IStep
   {
-    private Action<IStep> action;
-    private Predicate<IStep> canProcessPredicate;
+    private readonly Action<IStep> action;
+    private readonly Predicate<IStep> canProcessPredicate;
 
     public Step(Action<IStep> action, Predicate<IStep> canProcessPredicate = null)
     {
-      this.action = action;        
+      this.action = action;
       this.canProcessPredicate = canProcessPredicate ?? True;
     }
 
     public bool CanProcess()
     {
-        return this.canProcessPredicate(this);
+      return this.canProcessPredicate(this);
     }
 
     public bool IsAsync()
     {
-        return false;
+      return false;
     }
 
     public void Process()
     {
-      if(this.action != null)
-      {
-        this.action(this);
-      }
+      this.action?.Invoke(this);
     }
 
     public async Task ProcessAsync(CancellationToken token)
     {
-      #if !NET45
+#if !NET45
       await Task.FromException(new NotImplementedException());
-      #else
+#else
       var ctx = new TaskCompletionSource<bool>();
       ctx.SetException(new NotImplementedException());
       await ctx.Task;
-      #endif
+#endif
     }
 
     private bool True(IStep step)
     {
       return true;
-    }    
+    }
   }
 }