--- a/src/Anixe.TransactionSteps/PropertyBag.cs
+++ b/src/Anixe.TransactionSteps/PropertyBag.cs
@@ -9,86 +9,74 @@ namespace Anixe.TransactionSteps
     public override void Set<T>(T property)
     {
       var type = typeof(T);
-      this.typedProperties.Add(type, new PropertyBagItem<T>(property));
+      this.typedProperties.Add(type, property);
     }
 
     public override void Set<T>(string name, T property)
     {
-      this.namedProperties.Add(name, new PropertyBagItem<T>(property));
+      this.namedProperties.Add(name, property);
     }
   }
 
   public class PropertyBag : IPropertyBag
   {
-    protected class PropertyBagItem
-    {
-    }
-
-    protected class PropertyBagItem<T> : PropertyBagItem
-    {
-      private readonly T instance;
-      public PropertyBagItem(T instance)
-      {
-        this.instance = instance;
-      }
-
-      public T Value => this.instance;
-    }
-
-    protected readonly Dictionary<Type, PropertyBagItem> typedProperties;
-    protected readonly Dictionary<string, PropertyBagItem> namedProperties;
+    protected readonly Dictionary<Type, object> typedProperties;
+    protected readonly Dictionary<string, object> namedProperties;
 
     public PropertyBag()
     {
-      this.typedProperties = new Dictionary<Type, PropertyBagItem>();
-      this.namedProperties = new Dictionary<string, PropertyBagItem>();
+      this.typedProperties = new Dictionary<Type, object>();
+      this.namedProperties = new Dictionary<string, object>();
     }
 
-    private PropertyBag(Dictionary<Type, PropertyBagItem> typedProperties, Dictionary<string, PropertyBagItem> namedProperties)
+    private PropertyBag(IEnumerable<KeyValuePair<Type, object>> typedProperties, Dictionary<string, object> namedProperties)
     {
-      this.typedProperties = new Dictionary<Type, PropertyBagItem>(typedProperties);
-      this.namedProperties = new Dictionary<string, PropertyBagItem>(namedProperties);
+      this.namedProperties = new Dictionary<string, object>(namedProperties);
+      this.typedProperties = new Dictionary<Type, object>();
+      foreach (var prop in typedProperties)
+      {
+        this.typedProperties.Add(prop.Key, prop.Value);
+      }
     }
 
     public bool Contains<T>()
     {
-      return TryGet<T>() != null;
+      return this.typedProperties.ContainsKey(typeof(T));
     }
 
     public bool Contains<T>(string name)
     {
-      return TryGet<T>(name) != null;
+      return this.namedProperties.ContainsKey(name);
     }
 
     public virtual void Set<T>(T property)
     {
       var type = typeof(T);
-      this.typedProperties[type] = new PropertyBagItem<T>(property);
+      this.typedProperties[type] = property;
     }
 
     public virtual void Set<T>(string name, T property)
     {
-      this.namedProperties[name] = new PropertyBagItem<T>(property);
+      this.namedProperties[name] = property;
     }
 
     public T Get<T>()
     {
-      var retval = TryGet<T>();
-      if (retval != null)
+      var type = typeof(T);
+      if (this.typedProperties.TryGetValue(type, out var item))
       {
-        return retval.Value;
+        return (T)item;
       }
-      return default(T);
+      return default;
     }
 
     public T Get<T>(string name)
     {
-      var retval = TryGet<T>(name);
-      if (retval != null)
+      if (this.namedProperties.TryGetValue(name, out var item))
       {
-        return retval.Value;
+        return (T)item;
       }
-      return default(T);
+      return default;
     }
 
     public void Unset<T>()
@@ -104,30 +92,11 @@ namespace Anixe.TransactionSteps
 
     public IPropertyBag Clone(Type[] exclude)
     {
-      var cloned = this.typedProperties.Where(kvp => Array.Find(exclude, item => item == kvp.Key) == default).ToDictionary(g => g.Key, g => g.Value);
+      var cloned = this.typedProperties.Where(kvp => !Array.Exists(exclude, item => item == kvp.Key));
       return new PropertyBag(
         cloned,
         this.namedProperties
       );
-    }
-
-    private PropertyBagItem<T> TryGet<T>()
-    {
-      var type = typeof(T);
-      if (this.typedProperties.TryGetValue(type, out var item))
-      {
-        return (PropertyBagItem<T>)item;
-      }
-      return null;
-    }
-
-    private PropertyBagItem<T> TryGet<T>(string name)
-    {
-      if (this.namedProperties.TryGetValue(name, out var item))
-      {
-        return (PropertyBagItem<T>)item;
-      }
-      return null;
     }
   }
 }