--- a/src/Anixe.TransactionSteps/StepBase.cs
+++ b/src/Anixe.TransactionSteps/StepBase.cs
@@ -7,10 +7,7 @@ namespace Anixe.TransactionSteps
 {
   public abstract class AsyncStepBase<T> : StepBase<T> where T : class
   {
-    public bool IsAsync()
-    {
-      return true;
-    }
+    public bool IsAsync() => true;
     public void Process()
     {
       throw new NotImplementedException();
@@ -19,46 +16,41 @@ namespace Anixe.TransactionSteps
 
   public abstract class SyncStepBase<T> : StepBase<T> where T : class
   {
-    public bool IsAsync()
-    {
-      return false;
-    }
+    public bool IsAsync() => false;
     public Task ProcessAsync(CancellationToken token)
     {
       throw new NotImplementedException();
     }
   }
-  
+
   public abstract class StepBase<T> : StepBase where T : class
   {
-    public T Context
-    {
-      get;
-      set;
-    }
+    public T Context { get; set; }
   }
 
   public abstract class StepBase
   {
     private string stepName;
     public IServiceProvider Services { get; set; }
-    public string Name 
-    { 
-      get 
+
+    public string Name
+    {
+      get
       {
-        if(string.IsNullOrEmpty(this.stepName))
+        if (string.IsNullOrEmpty(this.stepName))
         {
           this.stepName = GetDefaultStepName();
         }
         return this.stepName;
-      } 
+      }
       set
       {
         this.stepName = value;
-      } 
+      }
     }
+
     public double TimeTaken { get; set; }
-    public int ProcessedItemsCount { get; set; }    
+    public int ProcessedItemsCount { get; set; }
     public LinkedList<IStep> Neighbourood { get; set; }
     public LinkedListNode<IStep> Current { get; set; }
     public bool WasFired { get; set; }
@@ -68,6 +60,6 @@ namespace Anixe.TransactionSteps
     private string GetDefaultStepName()
     {
       return this.GetType().Name;
-    }            
+    }
   }
 }