It looks that it is more efficient to use boxing of value types instead of always wrapping value with PropertyBagItem class. 

Before change:
```
|             Method |     Mean |     Error |    StdDev |   Median | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|------:|------:|------:|----------:|
|     Set_Value_Type | 3.976 us | 0.5293 us | 1.4754 us | 3.100 us |     - |     - |     - |     856 B |
| Set_Reference_Type | 4.553 us | 0.5825 us | 1.6713 us | 3.550 us |     - |     - |     - |     856 B |
|     Get_Value_Type | 4.290 us | 0.1211 us | 0.3455 us | 4.300 us |     - |     - |     - |         - |
| Get_Reference_Type | 4.687 us | 0.0993 us | 0.2361 us | 4.700 us |     - |     - |     - |         - |
```

After change:
```
|             Method |     Mean |     Error |    StdDev |   Median | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|------:|------:|------:|----------:|
|     Set_Value_Type | 5.022 us | 0.6291 us | 1.7640 us | 5.300 us |     - |     - |     - |     856 B |
| Set_Reference_Type | 3.356 us | 0.2597 us | 0.7367 us | 2.900 us |     - |     - |     - |     136 B |
|     Get_Value_Type | 3.904 us | 0.2187 us | 0.6169 us | 4.050 us |     - |     - |     - |         - |
| Get_Reference_Type | 2.597 us | 0.1166 us | 0.3030 us | 2.500 us |     - |     - |     - |         - |
```
Only setting of value types got worse but it is not common usage of PropertyBag - in most cases of usage it stores reference types.